Move data fetching functions to its proper place
