+++ frontend/src/infra/lib/validateToken.ts
